Extract statements_query_profile.go
